Refactor LeaveType query handling and add custom exceptions
Renamed query folders and classes for consistency and clarity. Added `NotFoundException` and `BadRequestException` to centralize error handling and improve maintainability. Updated related files to reflect the changes.

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -12,8 +13,13 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
 
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
-        //confirm it exists
+        //retrieve domain entity object
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(id: request.ID);
+        //confirm it exists
+        if (leaveTypeToDelete is not null)
+        {
+            throw new NotFoundException(nameof(Domain.LeaveType), request.ID);
+        }
 
         //delete it
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
